Fix schedulers' check-boxes after test ended and ui restailing
This commit fixes the following bug:
if you don't have a scheduler, the correspondig check box is
disabled, but when  after test ending the disabled check box is
enabled. Now when a test ends the function that confirgures the
schedulers' check-boxes is invoked.
This commit also changes the application name from S to S-GUI
and the application logo

Signed-off-by: Teodor Chetrusca <chetrusca63@gmail.com>
Acked-by: Luca Miccio <lucmiccio@gmail.com>