Fix a bug from file install.sh and fix menu-bar's bug
This commit fixes a bug of the file install.sh where the
dependency are checked. The script did not exit if some
dependencies were not satisfied. Now it ends when the script
check_dependency.sh returns an error.
This commit also fixes a bug that made disappear the menu-bar
on some distro.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>